Revert previous clumsy changes
