+++ projects/lgpt.vcproj
@@ -1053,10 +1053,6 @@
 						>
 					</File>
 					<File
-						RelativePath="..\sources\Application\utils\RandomNames.cpp"
-						>
-					</File>
-					<File
 						RelativePath="..\sources\Application\utils\HexBuffers.cpp"
 						>
 					</File>
