+++ projects/lgpt.vcxproj
@@ -301,7 +301,6 @@
     <ClCompile Include="..\sources\Application\utils\assert.cpp" />
     <ClCompile Include="..\sources\Application\utils\char.cpp" />
     <ClCompile Include="..\sources\Application\utils\fixed.cpp" />
-    <ClCompile Include="..\sources\Application\utils\RandomNames.cpp" />
     <ClCompile Include="..\sources\Application\utils\HexBuffers.cpp" />
     <ClCompile Include="..\sources\Application\Utils\wildcard.cpp" />
     <ClCompile Include="..\sources\Application\Player\Player.cpp" />
